requirements.txt
Django==5.2.7
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
groq==0.15.0
qrcode==7.4.2
Pillow==11.0.0
redis==5.0.1
django-redis==5.4.0
requests==2.31.0
python-dateutil==2.8.2
whitenoise==6.6.0
